Refactor ViewContactsFragment to Activity
